refactored logic to use users folder
