+++ src/main/java/users/admin/Client.java
@@ -1,3 +1,5 @@
+package users.admin;
+
 public class Client {
     private String ownerID;
     private String jobDuration;
